refactor: Move `WithErrorInfo` to `prqlc-ast`, use in `prqlc-parser`
A small refactor while working on other parts of the codebase this weekend
