+++ prqlc/prqlc-ast/Cargo.toml
@@ -13,6 +13,7 @@ doctest = false
 test = false
 
 [dependencies]
+anyhow = {version = "1.0", optional = true}
 enum-as-inner = "0.6.0"
 semver = {version = "1.0.20", features = ["serde"]}
 serde = {version = "1.0.193", features = ["derive"]}
@@ -21,3 +22,6 @@ strum = {version = "0.25.0", features = ["std", "derive"]}
 [dev-dependencies]
 insta = "1.34"
 serde_json = "1.0"
+
+[features]
+default = ["anyhow"]
